refactor(database): refactor migrations into separate versioned modules
- Split monolithic migrations.py into multiple migration version files under migration_versions
- Introduced base Migration class in migration_versions/base.py
- Added individual migration classes for each version (v001 to v005) in separate files
- Updated migrations.py to import and use these new migration classes
- Improved modularity and maintainability of migration code
- Added __init__.py and py.typed for migration_versions package

This refactor prepares the codebase for easier migration management and future extensions.
