+++ src/scriptrag/database/migrations.py
@@ -6,1592 +6,19 @@ and automatic schema upgrades.
 """
 
 import sqlite3
-from abc import ABC, abstractmethod
-from datetime import datetime
 from pathlib import Path
 from typing import Any
 
 from scriptrag.config import get_logger
 
+from .migration_versions.base import Migration
+from .migration_versions.v001_initial_schema import InitialSchemaMigration
+from .migration_versions.v002_vector_storage import VectorStorageMigration
+from .migration_versions.v003_fix_fts_columns import FixFTSColumnsMigration
+from .migration_versions.v004_scene_dependencies import SceneDependenciesMigration
+from .migration_versions.v005_script_bible import ScriptBibleMigration
+
 logger = get_logger(__name__)
-
-
-class Migration(ABC):
-    """Base class for database migrations."""
-
-    def __init__(self) -> None:
-        """Initialize migration."""
-        self.version: int = 0
-        self.description: str = ""
-        self.applied_at: datetime | None = None
-
-    @abstractmethod
-    def up(self, connection: sqlite3.Connection) -> None:
-        """Apply the migration.
-
-        Args:
-            connection: Database connection
-        """
-        pass
-
-    @abstractmethod
-    def down(self, connection: sqlite3.Connection) -> None:
-        """Rollback the migration.
-
-        Args:
-            connection: Database connection
-        """
-        pass
-
-    def __str__(self) -> str:
-        """String representation of migration."""
-        return f"Migration {self.version}: {self.description}"
-
-
-class InitialSchemaMigration(Migration):
-    """Initial schema creation migration."""
-
-    def __init__(self) -> None:
-        """Initialize initial schema migration."""
-        super().__init__()
-        self.version = 1
-        self.description = "Create initial database schema"
-
-    def up(self, connection: sqlite3.Connection) -> None:
-        """Create initial schema."""
-        logger.info("Applying initial schema migration")
-
-        # Schema version tracking
-        connection.execute(
-            """
-            CREATE TABLE IF NOT EXISTS schema_info (
-                version INTEGER PRIMARY KEY,
-                applied_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                description TEXT
-            )
-        """
-        )
-
-        # Core entity tables
-        connection.execute(
-            """
-            CREATE TABLE IF NOT EXISTS scripts (
-                id TEXT PRIMARY KEY,
-                title TEXT NOT NULL,
-                format TEXT DEFAULT 'screenplay',
-                author TEXT,
-                description TEXT,
-                genre TEXT,
-                logline TEXT,
-                fountain_source TEXT,
-                source_file TEXT,
-                is_series BOOLEAN DEFAULT FALSE,
-                title_page_json TEXT,
-                metadata_json TEXT,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-            )
-        """
-        )
-
-        connection.execute(
-            """
-            CREATE TABLE IF NOT EXISTS seasons (
-                id TEXT PRIMARY KEY,
-                script_id TEXT NOT NULL,
-                number INTEGER NOT NULL,
-                title TEXT,
-                description TEXT,
-                year INTEGER,
-                metadata_json TEXT,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                FOREIGN KEY (script_id) REFERENCES scripts(id) ON DELETE CASCADE,
-                UNIQUE(script_id, number)
-            )
-        """
-        )
-
-        connection.execute(
-            """
-            CREATE TABLE IF NOT EXISTS episodes (
-                id TEXT PRIMARY KEY,
-                script_id TEXT NOT NULL,
-                season_id TEXT,
-                number INTEGER NOT NULL,
-                title TEXT NOT NULL,
-                description TEXT,
-                air_date TIMESTAMP,
-                writer TEXT,
-                director TEXT,
-                metadata_json TEXT,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                FOREIGN KEY (script_id) REFERENCES scripts(id) ON DELETE CASCADE,
-                FOREIGN KEY (season_id) REFERENCES seasons(id) ON DELETE CASCADE
-            )
-        """
-        )
-
-        connection.execute(
-            """
-            CREATE TABLE IF NOT EXISTS characters (
-                id TEXT PRIMARY KEY,
-                script_id TEXT NOT NULL,
-                name TEXT NOT NULL,
-                description TEXT,
-                aliases_json TEXT,
-                metadata_json TEXT,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                FOREIGN KEY (script_id) REFERENCES scripts(id) ON DELETE CASCADE
-            )
-        """
-        )
-
-        connection.execute(
-            """
-            CREATE TABLE IF NOT EXISTS locations (
-                id TEXT PRIMARY KEY,
-                script_id TEXT NOT NULL,
-                interior BOOLEAN DEFAULT TRUE,
-                name TEXT NOT NULL,
-                time_of_day TEXT,
-                raw_text TEXT NOT NULL,
-                metadata_json TEXT,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                FOREIGN KEY (script_id) REFERENCES scripts(id) ON DELETE CASCADE
-            )
-        """
-        )
-
-        connection.execute(
-            """
-            CREATE TABLE IF NOT EXISTS scenes (
-                id TEXT PRIMARY KEY,
-                script_id TEXT NOT NULL,
-                episode_id TEXT,
-                season_id TEXT,
-                location_id TEXT,
-                heading TEXT,
-                description TEXT,
-                script_order INTEGER NOT NULL,
-                temporal_order INTEGER,
-                logical_order INTEGER,
-                estimated_duration_minutes REAL,
-                time_of_day TEXT,
-                date_in_story TEXT,
-                metadata_json TEXT,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                FOREIGN KEY (script_id) REFERENCES scripts(id) ON DELETE CASCADE,
-                FOREIGN KEY (episode_id) REFERENCES episodes(id) ON DELETE CASCADE,
-                FOREIGN KEY (season_id) REFERENCES seasons(id) ON DELETE CASCADE,
-                FOREIGN KEY (location_id) REFERENCES locations(id) ON DELETE SET NULL
-            )
-        """
-        )
-
-        connection.execute(
-            """
-            CREATE TABLE IF NOT EXISTS scene_elements (
-                id TEXT PRIMARY KEY,
-                scene_id TEXT NOT NULL,
-                element_type TEXT NOT NULL,
-                text TEXT NOT NULL,
-                raw_text TEXT NOT NULL,
-                order_in_scene INTEGER NOT NULL,
-                character_id TEXT,
-                character_name TEXT,
-                associated_dialogue_id TEXT,
-                metadata_json TEXT,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                FOREIGN KEY (scene_id) REFERENCES scenes(id) ON DELETE CASCADE,
-                FOREIGN KEY (character_id) REFERENCES characters(id) ON DELETE SET NULL,
-                FOREIGN KEY (associated_dialogue_id)
-                    REFERENCES scene_elements(id) ON DELETE SET NULL
-            )
-        """
-        )
-
-        # Graph database layer
-        connection.execute(
-            """
-            CREATE TABLE IF NOT EXISTS nodes (
-                id TEXT PRIMARY KEY,
-                node_type TEXT NOT NULL,
-                entity_id TEXT,
-                label TEXT,
-                properties_json TEXT,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-            )
-        """
-        )
-
-        connection.execute(
-            """
-            CREATE TABLE IF NOT EXISTS edges (
-                id TEXT PRIMARY KEY,
-                from_node_id TEXT NOT NULL,
-                to_node_id TEXT NOT NULL,
-                edge_type TEXT NOT NULL,
-                properties_json TEXT,
-                weight REAL DEFAULT 1.0,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                FOREIGN KEY (from_node_id) REFERENCES nodes(id) ON DELETE CASCADE,
-                FOREIGN KEY (to_node_id) REFERENCES nodes(id) ON DELETE CASCADE
-            )
-        """
-        )
-
-        # Embeddings for semantic search
-        connection.execute(
-            """
-            CREATE TABLE IF NOT EXISTS embeddings (
-                id TEXT PRIMARY KEY,
-                entity_type TEXT NOT NULL,
-                entity_id TEXT NOT NULL,
-                content TEXT NOT NULL,
-                embedding_model TEXT NOT NULL,
-                vector_json TEXT NOT NULL,
-                dimension INTEGER NOT NULL,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                UNIQUE(entity_type, entity_id, embedding_model)
-            )
-        """
-        )
-
-        # Create indexes
-        self._create_indexes(connection)
-
-        # Create triggers
-        self._create_triggers(connection)
-
-        # Create FTS tables
-        self._create_fts_tables(connection)
-
-        logger.info("Initial schema migration completed")
-
-    def down(self, connection: sqlite3.Connection) -> None:
-        """Drop all tables (destructive rollback)."""
-        logger.warning("Rolling back initial schema - this will drop all data!")
-
-        tables = [
-            "scene_elements_fts",
-            "characters_fts",
-            "embeddings",
-            "edges",
-            "nodes",
-            "scene_elements",
-            "scenes",
-            "locations",
-            "characters",
-            "episodes",
-            "seasons",
-            "scripts",
-            "schema_info",
-        ]
-
-        for table in tables:
-            connection.execute(f"DROP TABLE IF EXISTS {table}")
-
-        logger.info("Initial schema rollback completed")
-
-    def _create_indexes(self, connection: sqlite3.Connection) -> None:
-        """Create database indexes."""
-        indexes = [
-            "CREATE INDEX IF NOT EXISTS idx_scripts_title ON scripts(title)",
-            "CREATE INDEX IF NOT EXISTS idx_scripts_author ON scripts(author)",
-            "CREATE INDEX IF NOT EXISTS idx_scripts_genre ON scripts(genre)",
-            "CREATE INDEX IF NOT EXISTS idx_seasons_script_id ON seasons(script_id)",
-            """CREATE INDEX IF NOT EXISTS idx_seasons_number
-                ON seasons(script_id, number)""",
-            "CREATE INDEX IF NOT EXISTS idx_episodes_script_id ON episodes(script_id)",
-            "CREATE INDEX IF NOT EXISTS idx_episodes_season_id ON episodes(season_id)",
-            """CREATE INDEX IF NOT EXISTS idx_episodes_number
-                ON episodes(season_id, number)""",
-            """CREATE INDEX IF NOT EXISTS idx_characters_script_id
-                ON characters(script_id)""",
-            """CREATE INDEX IF NOT EXISTS idx_characters_name
-                ON characters(script_id, name)""",
-            """CREATE INDEX IF NOT EXISTS idx_locations_script_id
-                ON locations(script_id)""",
-            """CREATE INDEX IF NOT EXISTS idx_locations_name
-                ON locations(script_id, name)""",
-            "CREATE INDEX IF NOT EXISTS idx_scenes_script_id ON scenes(script_id)",
-            "CREATE INDEX IF NOT EXISTS idx_scenes_episode_id ON scenes(episode_id)",
-            "CREATE INDEX IF NOT EXISTS idx_scenes_season_id ON scenes(season_id)",
-            "CREATE INDEX IF NOT EXISTS idx_scenes_location_id ON scenes(location_id)",
-            """CREATE INDEX IF NOT EXISTS idx_scenes_script_order
-                ON scenes(script_id, script_order)""",
-            """CREATE INDEX IF NOT EXISTS idx_scenes_temporal_order
-                ON scenes(script_id, temporal_order)""",
-            """CREATE INDEX IF NOT EXISTS idx_scenes_logical_order
-                ON scenes(script_id, logical_order)""",
-            """CREATE INDEX IF NOT EXISTS idx_scene_elements_scene_id
-                ON scene_elements(scene_id)""",
-            """CREATE INDEX IF NOT EXISTS idx_scene_elements_character_id
-                ON scene_elements(character_id)""",
-            """CREATE INDEX IF NOT EXISTS idx_scene_elements_order
-                ON scene_elements(scene_id, order_in_scene)""",
-            """CREATE INDEX IF NOT EXISTS idx_scene_elements_type
-                ON scene_elements(element_type)""",
-            "CREATE INDEX IF NOT EXISTS idx_nodes_type ON nodes(node_type)",
-            "CREATE INDEX IF NOT EXISTS idx_nodes_entity_id ON nodes(entity_id)",
-            """CREATE INDEX IF NOT EXISTS idx_nodes_type_entity
-                ON nodes(node_type, entity_id)""",
-            "CREATE INDEX IF NOT EXISTS idx_edges_from_node ON edges(from_node_id)",
-            "CREATE INDEX IF NOT EXISTS idx_edges_to_node ON edges(to_node_id)",
-            "CREATE INDEX IF NOT EXISTS idx_edges_type ON edges(edge_type)",
-            """CREATE INDEX IF NOT EXISTS idx_edges_from_to
-                ON edges(from_node_id, to_node_id)""",
-            """CREATE INDEX IF NOT EXISTS idx_edges_type_from
-                ON edges(edge_type, from_node_id)""",
-            """CREATE INDEX IF NOT EXISTS idx_embeddings_entity
-                ON embeddings(entity_type, entity_id)""",
-            """CREATE INDEX IF NOT EXISTS idx_embeddings_model
-                ON embeddings(embedding_model)""",
-        ]
-
-        for index_sql in indexes:
-            connection.execute(index_sql)
-
-    def _create_triggers(self, connection: sqlite3.Connection) -> None:
-        """Create database triggers."""
-        triggers = [
-            """CREATE TRIGGER IF NOT EXISTS update_scripts_timestamp
-                AFTER UPDATE ON scripts
-                BEGIN
-                    UPDATE scripts SET updated_at = CURRENT_TIMESTAMP WHERE id = NEW.id;
-                END""",
-            """CREATE TRIGGER IF NOT EXISTS update_seasons_timestamp
-                AFTER UPDATE ON seasons
-                BEGIN
-                    UPDATE seasons SET updated_at = CURRENT_TIMESTAMP WHERE id = NEW.id;
-                END""",
-            """CREATE TRIGGER IF NOT EXISTS update_episodes_timestamp
-                AFTER UPDATE ON episodes
-                BEGIN
-                    UPDATE episodes SET updated_at = CURRENT_TIMESTAMP
-                    WHERE id = NEW.id;
-                END""",
-            """CREATE TRIGGER IF NOT EXISTS update_characters_timestamp
-                AFTER UPDATE ON characters
-                BEGIN
-                    UPDATE characters SET updated_at = CURRENT_TIMESTAMP
-                    WHERE id = NEW.id;
-                END""",
-            """CREATE TRIGGER IF NOT EXISTS update_locations_timestamp
-                AFTER UPDATE ON locations
-                BEGIN
-                    UPDATE locations SET updated_at = CURRENT_TIMESTAMP
-                    WHERE id = NEW.id;
-                END""",
-            """CREATE TRIGGER IF NOT EXISTS update_scenes_timestamp
-                AFTER UPDATE ON scenes
-                BEGIN
-                    UPDATE scenes SET updated_at = CURRENT_TIMESTAMP WHERE id = NEW.id;
-                END""",
-            """CREATE TRIGGER IF NOT EXISTS update_scene_elements_timestamp
-                AFTER UPDATE ON scene_elements
-                BEGIN
-                    UPDATE scene_elements SET updated_at = CURRENT_TIMESTAMP
-                    WHERE id = NEW.id;
-                END""",
-            """CREATE TRIGGER IF NOT EXISTS update_nodes_timestamp
-                AFTER UPDATE ON nodes
-                BEGIN
-                    UPDATE nodes SET updated_at = CURRENT_TIMESTAMP WHERE id = NEW.id;
-                END""",
-            """CREATE TRIGGER IF NOT EXISTS update_edges_timestamp
-                AFTER UPDATE ON edges
-                BEGIN
-                    UPDATE edges SET updated_at = CURRENT_TIMESTAMP WHERE id = NEW.id;
-                END""",
-        ]
-
-        for trigger_sql in triggers:
-            connection.execute(trigger_sql)
-
-    def _create_fts_tables(self, connection: sqlite3.Connection) -> None:
-        """Create full-text search tables and triggers."""
-        # Create FTS tables
-        connection.execute(
-            """
-            CREATE VIRTUAL TABLE IF NOT EXISTS scene_elements_fts USING fts5(
-                element_id,
-                text,
-                character_name,
-                scene_id,
-                content='scene_elements',
-                content_rowid='rowid'
-            )
-        """
-        )
-
-        connection.execute(
-            """
-            CREATE VIRTUAL TABLE IF NOT EXISTS characters_fts USING fts5(
-                character_id,
-                name,
-                description,
-                content='characters',
-                content_rowid='rowid'
-            )
-        """
-        )
-
-        # Create FTS triggers
-        fts_triggers = [
-            """CREATE TRIGGER IF NOT EXISTS scene_elements_fts_insert
-                AFTER INSERT ON scene_elements
-                BEGIN
-                    INSERT INTO scene_elements_fts(
-                        element_id, text, character_name, scene_id)
-                    VALUES (NEW.id, NEW.text, COALESCE(NEW.character_name, ''),
-                            NEW.scene_id);
-                END""",
-            """CREATE TRIGGER IF NOT EXISTS scene_elements_fts_update
-                AFTER UPDATE ON scene_elements
-                BEGIN
-                    UPDATE scene_elements_fts
-                    SET text = NEW.text,
-                        character_name = COALESCE(NEW.character_name, ''),
-                        scene_id = NEW.scene_id
-                    WHERE element_id = NEW.id;
-                END""",
-            """CREATE TRIGGER IF NOT EXISTS scene_elements_fts_delete
-                AFTER DELETE ON scene_elements
-                BEGIN
-                    DELETE FROM scene_elements_fts WHERE element_id = OLD.id;
-                END""",
-            """CREATE TRIGGER IF NOT EXISTS characters_fts_insert
-                AFTER INSERT ON characters
-                BEGIN
-                    INSERT INTO characters_fts(character_id, name, description)
-                    VALUES (NEW.id, NEW.name, COALESCE(NEW.description, ''));
-                END""",
-            """CREATE TRIGGER IF NOT EXISTS characters_fts_update
-                AFTER UPDATE ON characters
-                BEGIN
-                    UPDATE characters_fts
-                    SET name = NEW.name, description = COALESCE(NEW.description, '')
-                    WHERE character_id = NEW.id;
-                END""",
-            """CREATE TRIGGER IF NOT EXISTS characters_fts_delete
-                AFTER DELETE ON characters
-                BEGIN
-                    DELETE FROM characters_fts WHERE character_id = OLD.id;
-                END""",
-        ]
-
-        for trigger_sql in fts_triggers:
-            connection.execute(trigger_sql)
-
-
-class VectorStorageMigration(Migration):
-    """Migration to add sqlite-vec support to embeddings table."""
-
-    def __init__(self) -> None:
-        """Initialize vector storage migration."""
-        super().__init__()
-        self.version = 2
-        self.description = "Add sqlite-vec vector storage support to embeddings table"
-
-    def up(self, connection: sqlite3.Connection) -> None:
-        """Apply vector storage migration.
-
-        Adds vector_blob and vector_type columns to embeddings table
-        and converts existing JSON vectors to binary format.
-
-        Args:
-            connection: Database connection
-        """
-        cursor = connection.cursor()
-
-        try:
-            # Add new columns for sqlite-vec support
-            cursor.execute(
-                """
-                ALTER TABLE embeddings
-                ADD COLUMN vector_blob BLOB
-            """
-            )
-
-            cursor.execute(
-                """
-                ALTER TABLE embeddings
-                ADD COLUMN vector_type TEXT DEFAULT 'float32'
-            """
-            )
-
-            # Make vector_json nullable since we're adding vector_blob
-            cursor.execute(
-                """
-                CREATE TABLE embeddings_new (
-                    id TEXT PRIMARY KEY,
-                    entity_type TEXT NOT NULL,
-                    entity_id TEXT NOT NULL,
-                    content TEXT NOT NULL,
-                    embedding_model TEXT NOT NULL,
-                    vector_blob BLOB,
-                    vector_type TEXT DEFAULT 'float32',
-                    vector_json TEXT,
-                    dimension INTEGER NOT NULL,
-                    created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                    UNIQUE(entity_type, entity_id, embedding_model)
-                )
-            """
-            )
-
-            # Copy existing data
-            cursor.execute(
-                """
-                INSERT INTO embeddings_new (
-                    id, entity_type, entity_id, content, embedding_model,
-                    vector_json, dimension, created_at
-                )
-                SELECT
-                    id, entity_type, entity_id, content, embedding_model,
-                    vector_json, dimension, created_at
-                FROM embeddings
-            """
-            )
-
-            # Replace old table
-            cursor.execute("DROP TABLE embeddings")
-            cursor.execute("ALTER TABLE embeddings_new RENAME TO embeddings")
-
-            # Recreate index
-            cursor.execute(
-                """
-                CREATE INDEX IF NOT EXISTS idx_embeddings_entity
-                ON embeddings(entity_type, entity_id)
-            """
-            )
-            cursor.execute(
-                """
-                CREATE INDEX IF NOT EXISTS idx_embeddings_model
-                ON embeddings(embedding_model)
-            """
-            )
-
-            connection.commit()
-            logger.info("Vector storage migration applied successfully")
-
-        except Exception as e:
-            connection.rollback()
-            logger.error(f"Failed to apply vector storage migration: {e}")
-            raise
-
-    def down(self, connection: sqlite3.Connection) -> None:
-        """Rollback vector storage migration.
-
-        Removes vector_blob and vector_type columns from embeddings table.
-
-        Args:
-            connection: Database connection
-        """
-        cursor = connection.cursor()
-
-        try:
-            # Recreate original table structure
-            cursor.execute(
-                """
-                CREATE TABLE embeddings_rollback (
-                    id TEXT PRIMARY KEY,
-                    entity_type TEXT NOT NULL,
-                    entity_id TEXT NOT NULL,
-                    content TEXT NOT NULL,
-                    embedding_model TEXT NOT NULL,
-                    vector_json TEXT NOT NULL,
-                    dimension INTEGER NOT NULL,
-                    created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                    UNIQUE(entity_type, entity_id, embedding_model)
-                )
-            """
-            )
-
-            # Copy data back (only records with vector_json)
-            cursor.execute(
-                """
-                INSERT INTO embeddings_rollback (
-                    id, entity_type, entity_id, content, embedding_model,
-                    vector_json, dimension, created_at
-                )
-                SELECT
-                    id, entity_type, entity_id, content, embedding_model,
-                    vector_json, dimension, created_at
-                FROM embeddings
-                WHERE vector_json IS NOT NULL
-            """
-            )
-
-            # Replace table
-            cursor.execute("DROP TABLE embeddings")
-            cursor.execute("ALTER TABLE embeddings_rollback RENAME TO embeddings")
-
-            # Recreate indexes
-            cursor.execute(
-                """
-                CREATE INDEX IF NOT EXISTS idx_embeddings_entity
-                ON embeddings(entity_type, entity_id)
-            """
-            )
-            cursor.execute(
-                """
-                CREATE INDEX IF NOT EXISTS idx_embeddings_model
-                ON embeddings(embedding_model)
-            """
-            )
-
-            connection.commit()
-            logger.info("Vector storage migration rolled back successfully")
-
-        except Exception as e:
-            connection.rollback()
-            logger.error(f"Failed to rollback vector storage migration: {e}")
-            raise
-
-
-class FixFTSColumnsMigration(Migration):
-    """Fix FTS table column naming to match base tables."""
-
-    def __init__(self) -> None:
-        """Initialize FTS fix migration."""
-        super().__init__()
-        self.version = 3
-        self.description = "Fix FTS table column naming to match base tables"
-
-    def up(self, connection: sqlite3.Connection) -> None:
-        """Fix FTS column naming.
-
-        Args:
-            connection: Database connection
-        """
-        logger.info("Applying FTS column naming fix migration")
-
-        # Drop existing FTS tables and triggers
-        drop_statements = [
-            "DROP TRIGGER IF EXISTS scene_elements_fts_insert",
-            "DROP TRIGGER IF EXISTS scene_elements_fts_update",
-            "DROP TRIGGER IF EXISTS scene_elements_fts_delete",
-            "DROP TRIGGER IF EXISTS characters_fts_insert",
-            "DROP TRIGGER IF EXISTS characters_fts_update",
-            "DROP TRIGGER IF EXISTS characters_fts_delete",
-            "DROP TABLE IF EXISTS scene_elements_fts",
-            "DROP TABLE IF EXISTS characters_fts",
-        ]
-
-        for statement in drop_statements:
-            connection.execute(statement)
-
-        # Recreate FTS tables with correct column names
-        connection.execute(
-            """
-            CREATE VIRTUAL TABLE scene_elements_fts USING fts5(
-                id,
-                text,
-                character_name,
-                scene_id,
-                content='scene_elements',
-                content_rowid='rowid'
-            )
-            """
-        )
-
-        connection.execute(
-            """
-            CREATE VIRTUAL TABLE characters_fts USING fts5(
-                id,
-                name,
-                description,
-                content='characters',
-                content_rowid='rowid'
-            )
-            """
-        )
-
-        # Recreate triggers with correct column names
-        fts_triggers = [
-            """CREATE TRIGGER scene_elements_fts_insert
-                AFTER INSERT ON scene_elements
-                BEGIN
-                    INSERT INTO scene_elements_fts(id, text, character_name, scene_id)
-                    VALUES (NEW.id, NEW.text,
-                            COALESCE(NEW.character_name, ''), NEW.scene_id);
-                END""",
-            """CREATE TRIGGER scene_elements_fts_update
-                AFTER UPDATE ON scene_elements
-                BEGIN
-                    UPDATE scene_elements_fts
-                    SET text = NEW.text,
-                        character_name = COALESCE(NEW.character_name, ''),
-                        scene_id = NEW.scene_id
-                    WHERE id = NEW.id;
-                END""",
-            """CREATE TRIGGER scene_elements_fts_delete
-                AFTER DELETE ON scene_elements
-                BEGIN
-                    DELETE FROM scene_elements_fts WHERE id = OLD.id;
-                END""",
-            """CREATE TRIGGER characters_fts_insert
-                AFTER INSERT ON characters
-                BEGIN
-                    INSERT INTO characters_fts(id, name, description)
-                    VALUES (NEW.id, NEW.name, COALESCE(NEW.description, ''));
-                END""",
-            """CREATE TRIGGER characters_fts_update
-                AFTER UPDATE ON characters
-                BEGIN
-                    UPDATE characters_fts
-                    SET name = NEW.name,
-                        description = COALESCE(NEW.description, '')
-                    WHERE id = NEW.id;
-                END""",
-            """CREATE TRIGGER characters_fts_delete
-                AFTER DELETE ON characters
-                BEGIN
-                    DELETE FROM characters_fts WHERE id = OLD.id;
-                END""",
-        ]
-
-        for trigger in fts_triggers:
-            connection.execute(trigger)
-
-        logger.info("FTS column naming fix migration applied successfully")
-
-    def down(self, connection: sqlite3.Connection) -> None:
-        """Rollback FTS column fix.
-
-        Args:
-            connection: Database connection
-        """
-        logger.info("Rolling back FTS column naming fix migration")
-
-        try:
-            # Drop the fixed FTS tables and triggers
-            drop_statements = [
-                "DROP TRIGGER IF EXISTS scene_elements_fts_insert",
-                "DROP TRIGGER IF EXISTS scene_elements_fts_update",
-                "DROP TRIGGER IF EXISTS scene_elements_fts_delete",
-                "DROP TRIGGER IF EXISTS characters_fts_insert",
-                "DROP TRIGGER IF EXISTS characters_fts_update",
-                "DROP TRIGGER IF EXISTS characters_fts_delete",
-                "DROP TABLE IF EXISTS scene_elements_fts",
-                "DROP TABLE IF EXISTS characters_fts",
-            ]
-
-            for statement in drop_statements:
-                connection.execute(statement)
-
-            # Recreate the original (broken) FTS tables with element_id/character_id
-            connection.execute(
-                """
-                CREATE VIRTUAL TABLE scene_elements_fts USING fts5(
-                    element_id,
-                    text,
-                    character_name,
-                    scene_id,
-                    content='scene_elements',
-                    content_rowid='rowid'
-                )
-                """
-            )
-
-            connection.execute(
-                """
-                CREATE VIRTUAL TABLE characters_fts USING fts5(
-                    character_id,
-                    name,
-                    description,
-                    content='characters',
-                    content_rowid='rowid'
-                )
-                """
-            )
-
-            # Recreate original triggers
-            original_triggers = [
-                """CREATE TRIGGER scene_elements_fts_insert
-                    AFTER INSERT ON scene_elements
-                    BEGIN
-                        INSERT INTO scene_elements_fts(element_id, text,
-                                                     character_name, scene_id)
-                        VALUES (NEW.id, NEW.text,
-                                COALESCE(NEW.character_name, ''), NEW.scene_id);
-                    END""",
-                """CREATE TRIGGER scene_elements_fts_update
-                    AFTER UPDATE ON scene_elements
-                    BEGIN
-                        UPDATE scene_elements_fts
-                        SET text = NEW.text,
-                            character_name = COALESCE(NEW.character_name, ''),
-                            scene_id = NEW.scene_id
-                        WHERE element_id = NEW.id;
-                    END""",
-                """CREATE TRIGGER scene_elements_fts_delete
-                    AFTER DELETE ON scene_elements
-                    BEGIN
-                        DELETE FROM scene_elements_fts WHERE element_id = OLD.id;
-                    END""",
-                """CREATE TRIGGER characters_fts_insert
-                    AFTER INSERT ON characters
-                    BEGIN
-                        INSERT INTO characters_fts(character_id, name, description)
-                        VALUES (NEW.id, NEW.name, COALESCE(NEW.description, ''));
-                    END""",
-                """CREATE TRIGGER characters_fts_update
-                    AFTER UPDATE ON characters
-                    BEGIN
-                        UPDATE characters_fts
-                        SET name = NEW.name,
-                            description = COALESCE(NEW.description, '')
-                        WHERE character_id = NEW.id;
-                    END""",
-                """CREATE TRIGGER characters_fts_delete
-                    AFTER DELETE ON characters
-                    BEGIN
-                        DELETE FROM characters_fts WHERE character_id = OLD.id;
-                    END""",
-            ]
-
-            for trigger in original_triggers:
-                connection.execute(trigger)
-
-            connection.commit()
-            logger.info("FTS column naming fix migration rolled back successfully")
-
-        except Exception as e:
-            connection.rollback()
-            logger.error(f"Failed to rollback FTS column naming fix migration: {e}")
-            raise
-
-
-class SceneDependenciesMigration(Migration):
-    """Add scene dependencies table for logical ordering."""
-
-    def __init__(self) -> None:
-        """Initialize scene dependencies migration."""
-        super().__init__()
-        self.version = 4
-        self.description = "Add scene dependencies table for logical ordering"
-
-    def up(self, connection: sqlite3.Connection) -> None:
-        """Apply scene dependencies migration.
-
-        Args:
-            connection: Database connection
-        """
-        logger.info("Applying scene dependencies migration")
-
-        # Create scene dependencies table
-        connection.execute(
-            """
-            CREATE TABLE IF NOT EXISTS scene_dependencies (
-                id TEXT PRIMARY KEY,
-                from_scene_id TEXT NOT NULL,
-                to_scene_id TEXT NOT NULL,
-                dependency_type TEXT NOT NULL,
-                description TEXT,
-                strength REAL DEFAULT 1.0,
-                metadata_json TEXT,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                FOREIGN KEY (from_scene_id) REFERENCES scenes(id) ON DELETE CASCADE,
-                FOREIGN KEY (to_scene_id) REFERENCES scenes(id) ON DELETE CASCADE,
-                UNIQUE(from_scene_id, to_scene_id, dependency_type)
-            )
-            """
-        )
-
-        # Create indexes
-        connection.execute(
-            """
-            CREATE INDEX IF NOT EXISTS idx_scene_dependencies_from
-                ON scene_dependencies(from_scene_id)
-            """
-        )
-        connection.execute(
-            """
-            CREATE INDEX IF NOT EXISTS idx_scene_dependencies_to
-                ON scene_dependencies(to_scene_id)
-            """
-        )
-        connection.execute(
-            """
-            CREATE INDEX IF NOT EXISTS idx_scene_dependencies_type
-                ON scene_dependencies(dependency_type)
-            """
-        )
-
-        # Create trigger for updating timestamp
-        connection.execute(
-            """
-            CREATE TRIGGER IF NOT EXISTS update_scene_dependencies_timestamp
-                AFTER UPDATE ON scene_dependencies
-                BEGIN
-                    UPDATE scene_dependencies
-                    SET updated_at = CURRENT_TIMESTAMP WHERE id = NEW.id;
-                END
-            """
-        )
-
-        logger.info("Scene dependencies migration applied successfully")
-
-    def down(self, connection: sqlite3.Connection) -> None:
-        """Rollback scene dependencies migration.
-
-        Args:
-            connection: Database connection
-        """
-        logger.info("Rolling back scene dependencies migration")
-
-        # Drop trigger first
-        connection.execute("DROP TRIGGER IF EXISTS update_scene_dependencies_timestamp")
-
-        # Drop indexes
-        connection.execute("DROP INDEX IF EXISTS idx_scene_dependencies_from")
-        connection.execute("DROP INDEX IF EXISTS idx_scene_dependencies_to")
-        connection.execute("DROP INDEX IF EXISTS idx_scene_dependencies_type")
-
-        # Drop table
-        connection.execute("DROP TABLE IF EXISTS scene_dependencies")
-
-        logger.info("Scene dependencies migration rolled back successfully")
-
-
-class ScriptBibleMigration(Migration):
-    """Migration 5: Add Script Bible and continuity tracking tables."""
-
-    version = 5
-    description = "Add Script Bible and continuity tracking tables"
-
-    def up(self, connection: sqlite3.Connection) -> None:
-        """Apply Script Bible migration."""
-        logger.info("Applying Script Bible migration")
-
-        # Series information and bible metadata
-        connection.execute(
-            """
-            CREATE TABLE IF NOT EXISTS series_bibles (
-                id TEXT PRIMARY KEY,
-                script_id TEXT NOT NULL,
-                title TEXT NOT NULL,
-                description TEXT,
-                version INTEGER DEFAULT 1,
-                created_by TEXT,
-                status TEXT DEFAULT 'active',
-                bible_type TEXT DEFAULT 'series',
-                metadata_json TEXT,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                FOREIGN KEY (script_id) REFERENCES scripts(id) ON DELETE CASCADE
-            )
-            """
-        )
-
-        # Character development profiles and arcs
-        connection.execute(
-            """
-            CREATE TABLE IF NOT EXISTS character_profiles (
-                id TEXT PRIMARY KEY,
-                character_id TEXT NOT NULL,
-                script_id TEXT NOT NULL,
-                series_bible_id TEXT,
-                full_name TEXT,
-                age INTEGER,
-                occupation TEXT,
-                background TEXT,
-                personality_traits TEXT,
-                motivations TEXT,
-                fears TEXT,
-                goals TEXT,
-                physical_description TEXT,
-                distinguishing_features TEXT,
-                family_background TEXT,
-                relationship_status TEXT,
-                initial_state TEXT,
-                character_arc TEXT,
-                growth_trajectory TEXT,
-                first_appearance_episode_id TEXT,
-                last_appearance_episode_id TEXT,
-                total_appearances INTEGER DEFAULT 0,
-                notes TEXT,
-                metadata_json TEXT,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                FOREIGN KEY (character_id) REFERENCES characters(id) ON DELETE CASCADE,
-                FOREIGN KEY (script_id) REFERENCES scripts(id) ON DELETE CASCADE,
-                FOREIGN KEY (series_bible_id)
-                    REFERENCES series_bibles(id) ON DELETE CASCADE,
-                FOREIGN KEY (first_appearance_episode_id)
-                    REFERENCES episodes(id) ON DELETE SET NULL,
-                FOREIGN KEY (last_appearance_episode_id)
-                    REFERENCES episodes(id) ON DELETE SET NULL,
-                UNIQUE(character_id, script_id)
-            )
-            """
-        )
-
-        # World building and setting elements
-        connection.execute(
-            """
-            CREATE TABLE IF NOT EXISTS world_elements (
-                id TEXT PRIMARY KEY,
-                script_id TEXT NOT NULL,
-                series_bible_id TEXT,
-                element_type TEXT NOT NULL,
-                name TEXT NOT NULL,
-                category TEXT,
-                description TEXT,
-                rules_and_constraints TEXT,
-                visual_description TEXT,
-                first_introduced_episode_id TEXT,
-                first_introduced_scene_id TEXT,
-                usage_frequency INTEGER DEFAULT 0,
-                importance_level INTEGER DEFAULT 1,
-                related_locations_json TEXT,
-                related_characters_json TEXT,
-                continuity_notes TEXT,
-                established_rules_json TEXT,
-                notes TEXT,
-                metadata_json TEXT,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                FOREIGN KEY (script_id) REFERENCES scripts(id) ON DELETE CASCADE,
-                FOREIGN KEY (series_bible_id)
-                    REFERENCES series_bibles(id) ON DELETE CASCADE,
-                FOREIGN KEY (first_introduced_episode_id)
-                    REFERENCES episodes(id) ON DELETE SET NULL,
-                FOREIGN KEY (first_introduced_scene_id)
-                    REFERENCES scenes(id) ON DELETE SET NULL
-            )
-            """
-        )
-
-        # Timeline and chronology tracking
-        connection.execute(
-            """
-            CREATE TABLE IF NOT EXISTS story_timelines (
-                id TEXT PRIMARY KEY,
-                script_id TEXT NOT NULL,
-                series_bible_id TEXT,
-                name TEXT NOT NULL,
-                timeline_type TEXT DEFAULT 'main',
-                description TEXT,
-                start_date TEXT,
-                end_date TEXT,
-                duration_description TEXT,
-                reference_episodes_json TEXT,
-                reference_scenes_json TEXT,
-                notes TEXT,
-                metadata_json TEXT,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                FOREIGN KEY (script_id) REFERENCES scripts(id) ON DELETE CASCADE,
-                FOREIGN KEY (series_bible_id)
-                    REFERENCES series_bibles(id) ON DELETE CASCADE
-            )
-            """
-        )
-
-        # Timeline events for detailed chronology
-        connection.execute(
-            """
-            CREATE TABLE IF NOT EXISTS timeline_events (
-                id TEXT PRIMARY KEY,
-                timeline_id TEXT NOT NULL,
-                script_id TEXT NOT NULL,
-                event_name TEXT NOT NULL,
-                event_type TEXT DEFAULT 'plot',
-                story_date TEXT,
-                relative_order INTEGER,
-                duration_minutes INTEGER,
-                description TEXT,
-                episode_id TEXT,
-                scene_id TEXT,
-                related_characters_json TEXT,
-                establishes_json TEXT,
-                requires_json TEXT,
-                affects_json TEXT,
-                notes TEXT,
-                metadata_json TEXT,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                FOREIGN KEY (timeline_id)
-                    REFERENCES story_timelines(id) ON DELETE CASCADE,
-                FOREIGN KEY (script_id) REFERENCES scripts(id) ON DELETE CASCADE,
-                FOREIGN KEY (episode_id) REFERENCES episodes(id) ON DELETE SET NULL,
-                FOREIGN KEY (scene_id) REFERENCES scenes(id) ON DELETE SET NULL
-            )
-            """
-        )
-
-        # Production notes and tracking
-        connection.execute(
-            """
-            CREATE TABLE IF NOT EXISTS production_notes (
-                id TEXT PRIMARY KEY,
-                script_id TEXT NOT NULL,
-                series_bible_id TEXT,
-                note_type TEXT NOT NULL,
-                category TEXT,
-                title TEXT NOT NULL,
-                content TEXT NOT NULL,
-                priority INTEGER DEFAULT 3,
-                status TEXT DEFAULT 'open',
-                episode_id TEXT,
-                scene_id TEXT,
-                character_id TEXT,
-                assigned_to TEXT,
-                due_date TEXT,
-                resolved_date TEXT,
-                resolution_notes TEXT,
-                tags_json TEXT,
-                attachments_json TEXT,
-                metadata_json TEXT,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                created_by TEXT,
-                FOREIGN KEY (script_id) REFERENCES scripts(id) ON DELETE CASCADE,
-                FOREIGN KEY (series_bible_id)
-                    REFERENCES series_bibles(id) ON DELETE CASCADE,
-                FOREIGN KEY (episode_id) REFERENCES episodes(id) ON DELETE SET NULL,
-                FOREIGN KEY (scene_id) REFERENCES scenes(id) ON DELETE SET NULL,
-                FOREIGN KEY (character_id) REFERENCES characters(id) ON DELETE SET NULL
-            )
-            """
-        )
-
-        # Continuity issues tracking
-        connection.execute(
-            """
-            CREATE TABLE IF NOT EXISTS continuity_issues (
-                id TEXT PRIMARY KEY,
-                script_id TEXT NOT NULL,
-                issue_type TEXT NOT NULL,
-                severity TEXT DEFAULT 'warning',
-                status TEXT DEFAULT 'open',
-                description TEXT NOT NULL,
-                first_occurrence_episode_id TEXT,
-                first_occurrence_scene_id TEXT,
-                conflicting_episodes_json TEXT,
-                conflicting_scenes_json TEXT,
-                affected_characters_json TEXT,
-                affected_elements_json TEXT,
-                suggested_resolution TEXT,
-                resolution_notes TEXT,
-                resolved_by TEXT,
-                resolved_date TEXT,
-                metadata_json TEXT,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                FOREIGN KEY (script_id) REFERENCES scripts(id) ON DELETE CASCADE,
-                FOREIGN KEY (first_occurrence_episode_id)
-                    REFERENCES episodes(id) ON DELETE SET NULL,
-                FOREIGN KEY (first_occurrence_scene_id)
-                    REFERENCES scenes(id) ON DELETE SET NULL
-            )
-            """
-        )
-
-        # Continuity validations log
-        connection.execute(
-            """
-            CREATE TABLE IF NOT EXISTS continuity_validations (
-                id TEXT PRIMARY KEY,
-                script_id TEXT NOT NULL,
-                validation_type TEXT NOT NULL,
-                validation_date TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                episodes_validated INTEGER DEFAULT 0,
-                scenes_validated INTEGER DEFAULT 0,
-                issues_found INTEGER DEFAULT 0,
-                issues_resolved INTEGER DEFAULT 0,
-                validation_rules_json TEXT,
-                results_summary_json TEXT,
-                metadata_json TEXT,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                FOREIGN KEY (script_id) REFERENCES scripts(id) ON DELETE CASCADE
-            )
-            """
-        )
-
-        # Continuity notes table
-        connection.execute(
-            """
-            CREATE TABLE IF NOT EXISTS continuity_notes (
-                id TEXT PRIMARY KEY,
-                script_id TEXT NOT NULL,
-                series_bible_id TEXT,
-                note_type TEXT NOT NULL,
-                severity TEXT DEFAULT 'medium',
-                status TEXT DEFAULT 'open',
-                title TEXT NOT NULL,
-                description TEXT NOT NULL,
-                suggested_resolution TEXT,
-                episode_id TEXT,
-                scene_id TEXT,
-                character_id TEXT,
-                world_element_id TEXT,
-                timeline_event_id TEXT,
-                related_episodes_json TEXT,
-                related_scenes_json TEXT,
-                related_characters_json TEXT,
-                reported_by TEXT,
-                assigned_to TEXT,
-                resolution_notes TEXT,
-                resolved_at TIMESTAMP,
-                tags_json TEXT,
-                metadata_json TEXT,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                FOREIGN KEY (script_id) REFERENCES scripts(id) ON DELETE CASCADE,
-                FOREIGN KEY (series_bible_id)
-                    REFERENCES series_bibles(id) ON DELETE CASCADE,
-                FOREIGN KEY (episode_id) REFERENCES episodes(id) ON DELETE SET NULL,
-                FOREIGN KEY (scene_id) REFERENCES scenes(id) ON DELETE SET NULL,
-                FOREIGN KEY (character_id) REFERENCES characters(id) ON DELETE SET NULL,
-                FOREIGN KEY (world_element_id)
-                    REFERENCES world_elements(id) ON DELETE SET NULL,
-                FOREIGN KEY (timeline_event_id)
-                    REFERENCES timeline_events(id) ON DELETE SET NULL
-            )
-            """
-        )
-
-        # Character knowledge tracking
-        connection.execute(
-            """
-            CREATE TABLE IF NOT EXISTS character_knowledge (
-                id TEXT PRIMARY KEY,
-                character_id TEXT NOT NULL,
-                script_id TEXT NOT NULL,
-                knowledge_type TEXT NOT NULL,
-                knowledge_subject TEXT NOT NULL,
-                knowledge_description TEXT,
-                acquired_episode_id TEXT,
-                acquired_scene_id TEXT,
-                acquisition_method TEXT,
-                first_used_episode_id TEXT,
-                first_used_scene_id TEXT,
-                usage_count INTEGER DEFAULT 0,
-                should_know_before TEXT,
-                verification_status TEXT DEFAULT 'unverified',
-                confidence_level REAL DEFAULT 1.0,
-                notes TEXT,
-                metadata_json TEXT,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                FOREIGN KEY (character_id) REFERENCES characters(id) ON DELETE CASCADE,
-                FOREIGN KEY (script_id) REFERENCES scripts(id) ON DELETE CASCADE,
-                FOREIGN KEY (acquired_episode_id)
-                    REFERENCES episodes(id) ON DELETE SET NULL,
-                FOREIGN KEY (acquired_scene_id)
-                    REFERENCES scenes(id) ON DELETE SET NULL,
-                FOREIGN KEY (first_used_episode_id)
-                    REFERENCES episodes(id) ON DELETE SET NULL,
-                FOREIGN KEY (first_used_scene_id)
-                    REFERENCES scenes(id) ON DELETE SET NULL
-            )
-            """
-        )
-
-        # Plot threads tracking
-        connection.execute(
-            """
-            CREATE TABLE IF NOT EXISTS plot_threads (
-                id TEXT PRIMARY KEY,
-                script_id TEXT NOT NULL,
-                series_bible_id TEXT,
-                name TEXT NOT NULL,
-                thread_type TEXT DEFAULT 'main',
-                priority INTEGER DEFAULT 1,
-                description TEXT,
-                initial_setup TEXT,
-                central_conflict TEXT,
-                resolution TEXT,
-                status TEXT DEFAULT 'active',
-                introduced_episode_id TEXT,
-                resolved_episode_id TEXT,
-                total_episodes_involved INTEGER DEFAULT 0,
-                primary_characters_json TEXT,
-                supporting_characters_json TEXT,
-                key_scenes_json TEXT,
-                resolution_scenes_json TEXT,
-                notes TEXT,
-                metadata_json TEXT,
-                created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                FOREIGN KEY (script_id) REFERENCES scripts(id) ON DELETE CASCADE,
-                FOREIGN KEY (series_bible_id)
-                    REFERENCES series_bibles(id) ON DELETE CASCADE,
-                FOREIGN KEY (introduced_episode_id)
-                    REFERENCES episodes(id) ON DELETE SET NULL,
-                FOREIGN KEY (resolved_episode_id)
-                    REFERENCES episodes(id) ON DELETE SET NULL
-            )
-            """
-        )
-
-        # Create indexes for Script Bible tables
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_series_bibles_script_id
-                ON series_bibles(script_id)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_series_bibles_status
-                ON series_bibles(status)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_series_bibles_type
-                ON series_bibles(bible_type)"""
-        )
-
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_character_profiles_character_id
-                ON character_profiles(character_id)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_character_profiles_script_id
-                ON character_profiles(script_id)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_character_profiles_series_bible_id
-                ON character_profiles(series_bible_id)"""
-        )
-
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_world_elements_script_id
-                ON world_elements(script_id)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_world_elements_series_bible_id
-                ON world_elements(series_bible_id)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_world_elements_type
-                ON world_elements(element_type)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_world_elements_name
-                ON world_elements(name)"""
-        )
-
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_story_timelines_script_id
-                ON story_timelines(script_id)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_story_timelines_series_bible_id
-                ON story_timelines(series_bible_id)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_story_timelines_type
-                ON story_timelines(timeline_type)"""
-        )
-
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_timeline_events_timeline_id
-                ON timeline_events(timeline_id)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_timeline_events_script_id
-                ON timeline_events(script_id)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_timeline_events_episode_id
-                ON timeline_events(episode_id)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_timeline_events_scene_id
-                ON timeline_events(scene_id)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_timeline_events_type
-                ON timeline_events(event_type)"""
-        )
-
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_production_notes_script_id
-                ON production_notes(script_id)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_production_notes_series_bible_id
-                ON production_notes(series_bible_id)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_production_notes_type
-                ON production_notes(note_type)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_production_notes_status
-                ON production_notes(status)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_production_notes_priority
-                ON production_notes(priority)"""
-        )
-
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_continuity_issues_script_id
-                ON continuity_issues(script_id)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_continuity_issues_type
-                ON continuity_issues(issue_type)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_continuity_issues_severity
-                ON continuity_issues(severity)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_continuity_issues_status
-                ON continuity_issues(status)"""
-        )
-
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_continuity_validations_script_id
-                ON continuity_validations(script_id)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_continuity_validations_type
-                ON continuity_validations(validation_type)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_continuity_validations_date
-                ON continuity_validations(validation_date)"""
-        )
-
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_continuity_notes_script_id
-                ON continuity_notes(script_id)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_continuity_notes_bible_id
-                ON continuity_notes(series_bible_id)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_continuity_notes_type
-                ON continuity_notes(note_type)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_continuity_notes_status
-                ON continuity_notes(status)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_continuity_notes_severity
-                ON continuity_notes(severity)"""
-        )
-
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_character_knowledge_character_id
-                ON character_knowledge(character_id)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_character_knowledge_script_id
-                ON character_knowledge(script_id)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_character_knowledge_type
-                ON character_knowledge(knowledge_type)"""
-        )
-
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_plot_threads_script_id
-                ON plot_threads(script_id)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_plot_threads_bible_id
-                ON plot_threads(series_bible_id)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_plot_threads_type
-                ON plot_threads(thread_type)"""
-        )
-        connection.execute(
-            """CREATE INDEX IF NOT EXISTS idx_plot_threads_status
-                ON plot_threads(status)"""
-        )
-
-        logger.info("Script Bible migration applied successfully")
-
-    def down(self, connection: sqlite3.Connection) -> None:
-        """Rollback Script Bible migration."""
-        logger.warning("Rolling back Script Bible migration")
-
-        # Drop indexes
-        indexes = [
-            "idx_series_bibles_script_id",
-            "idx_series_bibles_status",
-            "idx_series_bibles_type",
-            "idx_character_profiles_character_id",
-            "idx_character_profiles_script_id",
-            "idx_character_profiles_series_bible_id",
-            "idx_world_elements_script_id",
-            "idx_world_elements_series_bible_id",
-            "idx_world_elements_type",
-            "idx_world_elements_name",
-            "idx_story_timelines_script_id",
-            "idx_story_timelines_series_bible_id",
-            "idx_story_timelines_type",
-            "idx_timeline_events_timeline_id",
-            "idx_timeline_events_script_id",
-            "idx_timeline_events_episode_id",
-            "idx_timeline_events_scene_id",
-            "idx_timeline_events_type",
-            "idx_production_notes_script_id",
-            "idx_production_notes_series_bible_id",
-            "idx_production_notes_type",
-            "idx_production_notes_status",
-            "idx_production_notes_priority",
-            "idx_continuity_issues_script_id",
-            "idx_continuity_issues_type",
-            "idx_continuity_issues_severity",
-            "idx_continuity_issues_status",
-            "idx_continuity_validations_script_id",
-            "idx_continuity_validations_type",
-            "idx_continuity_validations_date",
-            "idx_continuity_notes_script_id",
-            "idx_continuity_notes_bible_id",
-            "idx_continuity_notes_type",
-            "idx_continuity_notes_status",
-            "idx_continuity_notes_severity",
-            "idx_character_knowledge_character_id",
-            "idx_character_knowledge_script_id",
-            "idx_character_knowledge_type",
-            "idx_plot_threads_script_id",
-            "idx_plot_threads_bible_id",
-            "idx_plot_threads_type",
-            "idx_plot_threads_status",
-        ]
-
-        for idx in indexes:
-            connection.execute(f"DROP INDEX IF EXISTS {idx}")
-
-        # Drop tables in reverse order of creation
-        tables = [
-            "plot_threads",
-            "character_knowledge",
-            "continuity_notes",
-            "continuity_validations",
-            "continuity_issues",
-            "production_notes",
-            "timeline_events",
-            "story_timelines",
-            "world_elements",
-            "character_profiles",
-            "series_bibles",
-        ]
-
-        for table in tables:
-            connection.execute(f"DROP TABLE IF EXISTS {table}")
-
-        logger.info("Script Bible migration rolled back successfully")
 
 
 class MigrationRunner:
@@ -1894,3 +321,6 @@ def migrate_database(db_path: str | Path, target_version: int | None = None) -> 
     if target_version is None:
         return runner.migrate_to_latest()
     return runner.migrate_to_version(target_version)
+
+
+__all__ = ["Migration", "MigrationRunner", "initialize_database", "migrate_database"]
